feat: :sparkles: Experiments with DataURI encoding for video files
Using this system, it is possible to keep the videos and the game logic ina  single HTML file, which can make distributions even easier. This is a very brittle test, but I think it was instructive to see the limitations and difficulties. The large files resulting from the text encoding make editing and loading the HTML quite unresponsive. Still, interesting.